change sales and po count to numeric type

+++ models/sales/sales_clean_model.sql
@@ -31,7 +31,7 @@ select
   dist_kode_outlet as dist_kode_outlet,
   dist_outlet_name as dist_outlet_name,
   safe_cast(regexp_replace(sales_pcs, r'[^0-9.]', '') as int64) as sales_pcs,
-  safe_cast(regexp_replace(sales_ctn, r'[^0-9.]', '') as int64) as sales_ctn,
+  safe_cast(regexp_replace(sales_ctn, r'[^0-9.]', '') as numeric) as sales_ctn,
   safe_cast(regexp_replace(sales_qty3, r'[^0-9.]', '') as int64) as sales_qty3,
   safe_cast(price_per_pc_rev as numeric) as price_per_pc_rev,
   safe_cast(regexp_replace(sales_value_gtv, r'[^0-9.]', '') as numeric) as sales_value_gtv,
@@ -43,7 +43,7 @@ select
   store_name_clean as store_name_clean,
   parse_date('%d/%m/%E4Y', replace(delivery_date, '-', '/')) as delivery_date,
   safe_cast(regexp_replace(po_pcs, r'[^0-9.]', '') as int64) as po_pcs,
-  safe_cast(regexp_replace(po_ctn, r'[^0-9.]', '') as int64) as po_ctn,
+  safe_cast(regexp_replace(po_ctn, r'[^0-9.]', '') as numeric) as po_ctn,
   safe_cast(regexp_replace(po_value_rbp, r'[^0-9.]', '') as numeric) as po_value_rbp
 from {{ ref('sales_backup_model') }}
 where backup_date = {{ get_workflow_date() }}
